+++ R/unload.r
@@ -67,9 +67,9 @@ reload = function (mod) {
     call_hook(mod_ns, '.on_unload', mod_ns)
     deregister_mod(info)
 
-    for (import_ns in namespace_info(mod_ns, 'imports')) {
-        call_hook(import_ns, '.on_unload', import_ns)
-        deregister_mod(attr(import_ns, 'info'))
+    for (import in namespace_info(mod_ns, 'imports')) {
+        call_hook(import$ns, '.on_unload', import$ns)
+        deregister_mod(import$info)
     }
 
     on.exit({