+++ tests/testthat/mod/issue211/__init__.r
@@ -1,0 +1,3 @@
+#' @export
+box::use(./sub/a)
+box::use(./sub/b)